requirements.txt
# AI
google-genai
openai

# Crawling
playwright
playwright-stealth

# Other
setuptools # required by playwright-stealth
retry